requirements.txt
streamlit
httpx
python-dotenv
PyMuPDF
python-docx